+++ src/fsd/1-pages/plan-survival/tabs/milestones-tab.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { MiscIcon } from '@/fsd/5-shared/ui/icons';
-import { TieredRewardCell, TieredRewardGrid } from '@/fsd/5-shared/ui/tiered-reward-grid';
+import { buildProgressRewardCells, TieredRewardCell, TieredRewardGrid } from '@/fsd/5-shared/ui/tiered-reward-grid';
 
 import { getShopCurrencyIconKey, getShopCurrencyLabel } from '@/fsd/4-entities/shops';
 import { ISurvivalEvent, survivalRewardInfo } from '@/fsd/4-entities/survival';
@@ -11,18 +11,11 @@ interface Props {
 }
 
 export const MilestonesTab: React.FC<Props> = ({ event }) => {
-    const milestoneCells: TieredRewardCell[] = event.milestoneRewards.map((milestone, index) => {
-        const rewardIds = milestone.reward
-            ? [milestone.reward]
-            : (milestone.chest?.flatMap(bundle => bundle.rewards) ?? []);
-        const resolved = rewardIds.map(rewardId => survivalRewardInfo(rewardId));
-        return {
-            key: index,
-            title: resolved.length === 1 ? resolved[0].label : undefined,
-            rewards: resolved.map(({ icon, qty }) => ({ icon, qty })),
-            costLabel: `${milestone.requiredProgress.toLocaleString()} pts`,
-        };
-    });
+    const milestoneCells = buildProgressRewardCells(
+        event.milestoneRewards,
+        survivalRewardInfo,
+        milestone => `${milestone.requiredProgress.toLocaleString()} pts`
+    );
 
     const currencyIconKey = getShopCurrencyIconKey(event.resourceId);
     const currencyLabel = getShopCurrencyLabel(event.resourceId);
