+++ src/fsd/5-shared/ui/tab-bar/index.ts
@@ -1,0 +1,1 @@
+export { TabBar } from './tab-bar';
