+++ src/fsd/5-shared/ui/tiered-reward-grid/index.ts
@@ -1,3 +1,5 @@
+export { buildProgressRewardCells } from './build-progress-reward-cells';
+export type { ProgressRewardEntry, ResolvedReward } from './build-progress-reward-cells';
 export { TieredRewardGrid } from './tiered-reward-grid';
 export { TIERED_REWARD_ICON_SIZE } from './tiered-reward-grid.model';
 export type { TieredRewardCell, TieredRewardCellReward } from './tiered-reward-grid.model';
